refactor(tests): Merge some relevant test files and add some additional test methods
