+++ tests/Unit/Loggers/PrettyLoggerTest.php
@@ -253,4 +253,160 @@ class PrettyLoggerTest extends TestCase
         $this->assertFileContains($filePath, '#extension: null');
         $this->assertFileContains($filePath, '#silent: false');
     }
+
+    public function testCanLogEmergencyLogLevelInPrettyLog()
+    {
+        $data = 'emergency';
+
+        $this->logger->emergency($data);
+
+        $filePath = $this->getLogFilePath('emergency.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogAlertLogLevelInPrettyLog()
+    {
+        $data = 'alert';
+
+        $this->logger->alert($data);
+
+        $filePath = $this->getLogFilePath('alert.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogCriticalLogLevelInPrettyLog()
+    {
+        $data = 'critical';
+
+        $this->logger->critical($data);
+
+        $filePath = $this->getLogFilePath('critical.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogErrorLogLevelInPrettyLog()
+    {
+        $data = 'error';
+
+        $this->logger->error($data);
+
+        $filePath = $this->getLogFilePath('error.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogWarningLogLevelInPrettyLog()
+    {
+        $data = 'warning';
+
+        $this->logger->warning($data);
+
+        $filePath = $this->getLogFilePath('warning.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogNoticeLogLevelInPrettyLog()
+    {
+        $data = 'notice';
+
+        $this->logger->notice($data);
+
+        $filePath = $this->getLogFilePath('notice.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogInfoLogLevelInPrettyLog()
+    {
+        $data = 'info';
+
+        $this->logger->info($data);
+
+        $filePath = $this->getLogFilePath('info.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogDebugLogLevelInPrettyLog()
+    {
+        $data = 'debug';
+
+        $this->logger->debug($data);
+
+        $filePath = $this->getLogFilePath('debug.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
+
+    public function testCanLogExceptionWithoutTraceInPrettyLog()
+    {
+        try {
+            throw new \Exception(
+                'exception message',
+                '500'
+            );
+        } catch (\Throwable $e) {
+            $line = $e->getLine();
+            $this->logger->exception($e);
+        }
+
+        $filePath = $this->getLogFilePath('exception.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, 'array:5');
+        $this->assertFileContains($filePath, '"class" => "Exception"');
+        $this->assertFileContains($filePath, '"massage" => "exception message"');
+        $this->assertFileContains($filePath, '"code" => 500');
+        $this->assertFileContains($filePath, '"file"');
+        $this->assertFileContains($filePath, '"line" => ' . ($line ?? 0));
+    }
+
+    public function testCanLogExceptionWithTraceInPrettyLog()
+    {
+        try {
+            throw new \Exception(
+                'exception message',
+                '500'
+            );
+        } catch (\Throwable $e) {
+            $line = $e->getLine();
+            $this->logger->exception($e, true);
+        }
+
+        $filePath = $this->getLogFilePath('exception.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, 'array:6');
+        $this->assertFileContains($filePath, '"class" => "Exception"');
+        $this->assertFileContains($filePath, '"massage" => "exception message"');
+        $this->assertFileContains($filePath, '"code" => 500');
+        $this->assertFileContains($filePath, '"file"');
+        $this->assertFileContains($filePath, '"line" => ' . ($line ?? 0));
+        $this->assertFileContains($filePath, '"trace"');
+    }
+
+    public function testCanLogDataWithCustomLevelInPrettyLog()
+    {
+        $data = 'test';
+
+        $this->logger->log($data, 'custom-level');
+
+        $filePath = $this->getLogFilePath('custom-level.log');
+
+        $this->assertFileExists($filePath);
+        $this->assertFileContains($filePath, $data);
+    }
 }
